Refactored InspectItemEvent into discrete file
